% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contour_detector.R
\name{image_contour_detector.SpatRaster}
\alias{image_contour_detector.SpatRaster}
\title{Unsupervised Smooth Contour Lines Detection for SpatRaster objects}
\usage{
\method{image_contour_detector}{SpatRaster}(x, Q = 2, as_sf = FALSE, ...)
}
\arguments{
\item{x}{a SpatRaster object}

\item{Q}{numeric value with the pixel quantization step}

\item{as_sf}{Boolean. Set to TRUE to export lines as sf spatial objects}

\item{...}{further arguments passed on to \code{image_contour_detector.matrix}}
}
\value{
In case \code{as_sf} is \code{FALSE}: an object of class cld which as described in \code{\link{image_contour_detector}}\cr
In case \code{as_sf} is \code{TRUE}: an object of class sf with columns ss and length_m
}
\description{
Unsupervised Smooth Contour Lines Detection for SpatRaster objects
}
\examples{
\dontshow{
if(require(terra) && require(sf))
\{
}
\donttest{
library(terra)
x   <- rast(system.file("extdata", "landscape.tif", package="image.ContourDetector"))

contourlines <- image_contour_detector(x)
image(x)
plot(contourlines, add = TRUE, col = "blue", lwd = 10)

contourlines <- image_contour_detector(x, as_sf = TRUE)
}
\dontshow{
\}
# End of main if statement running only if the required packages are installed
}
}
\seealso{
\code{\link{image_contour_detector}}
}
