% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contour_detector.R
\name{image_contour_detector.SpatRaster}
\alias{image_contour_detector.SpatRaster}
\title{Unsupervised Smooth Contour Lines Detection for SpatRaster objects}
\usage{
\method{image_contour_detector}{SpatRaster}(x, Q = 2, as_sf = FALSE, ...)
}
\arguments{
\item{x}{a SpatRaster object}

\item{Q}{numeric value with the pixel quantization step}

\item{as_sf}{Boolean. Set to TRUE to export lines as sf spatial objects}

\item{...}{further arguments passed on to \code{image_contour_detector.matrix}}
}
\value{
an object of class cld which as described in \code{\link{image_contour_detector}}
}
\description{
Unsupervised Smooth Contour Lines Detection for SpatRaster objects
}
\seealso{
\code{\link{image_contour_detector}}
}
